replies-pipeline/requirements.txt
elasticsearch==7.10.0
kafka-python==2.0.2
matplotlib==3.3.3
#nltk==3.5
numpy==1.19.4
#spacy==2.3.5
# pandas==1.1.4
# pickleshare==0.7.5
#pqdict==1.1.1
psycopg2-binary==2.8.6
scikit-learn==0.23.2
scipy==1.5.4
#tensorflow-cpu==2.4.1
#transformers==4.1.1
#geopy==2.1.0
# vaderSentiment==3.3.2
